+++ pallets/programs/src/benchmarking.rs
@@ -53,7 +53,7 @@ benchmarks! {
 
     let value = CurrencyOf::<T>::minimum_balance().saturating_mul(1_000_000_000u32.into());
     let _ = CurrencyOf::<T>::make_free_balance_be(&program_modification_account, value);
-    <Bytecode<T>>::insert(program_hash.clone(), ProgramInfo {bytecode: program, program_modification_account: program_modification_account.clone()});
+    <Programs<T>>::insert(program_hash.clone(), ProgramInfo {bytecode: program, program_modification_account: program_modification_account.clone()});
     let mut program_hashes = vec![program_hash.clone()];
     for _ in 1..p {
        program_hashes.push(random_hash);