+++ pallets/propagation/src/tests.rs
@@ -61,7 +61,7 @@ fn knows_how_to_mock_several_http_calls() {
         Propagation::post_dkg(1).unwrap();
 
         System::set_block_number(3);
-        pallet_programs::Bytecode::<Test>::insert(
+        pallet_programs::Programs::<Test>::insert(
             <Test as frame_system::Config>::Hash::default(),
             ProgramInfo { bytecode: vec![], program_modification_account: 1 },
         );