+++ CHANGELOG.md
@@ -11,12 +11,13 @@ At the moment this project **does not** adhere to
 
 -  Test CLI which calls the same code as in integration tests ([#417](https://github.com/entropyxyz/entropy-core/pull/417))
 
-### Added
-- Pointer for Programs ([#536](https://github.com/entropyxyz/entropy-core/pull/536/))
-
 ### Breaking Changes
 
 - In [#536](https://github.com/entropyxyz/entropy-core/pull/536/), the registered struct no longer holds a program but rather a hash of a program that is set in the set_program function 
+
+### Added
+- Pointer for Programs ([#536](https://github.com/entropyxyz/entropy-core/pull/536/))
+
 
 ## [0.0.9](https://github.com/entropyxyz/entropy-core/compare/release/v0.0.8..release/v0.0.9) - 2023-11-30